# docs/apis/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bifrost Gateway",
    "version": "1.3",
    "description": "The Bifrost API provides a unified interface to various AI models and services.",
    "contact": {
      "name": "Contact Us",
      "url": "https://getmaxim.ai/bifrost"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/logs": {
      "get": {
        "summary": "Search and filter logs",
        "operationId": "searchLogs",
        "tags": [
          "Logging"
        ],
        "parameters": [
          {
            "name": "providers",
            "in": "query",
            "description": "Comma-separated list of providers to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "models",
            "in": "query",
            "description": "Comma-separated list of models to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Comma-separated list of statuses to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objects",
            "in": "query",
            "description": "Comma-separated list of objects to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "Start time for the time range filter (RFC3339 format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "End time for the time range filter (RFC3339 format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "min_latency",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_latency",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_tokens",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "max_tokens",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "min_cost",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_cost",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "content_search",
            "in": "query",
            "description": "Full-text search in request/response bodies",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of logs to return",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of logs to skip",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string",
              "enum": [
                "timestamp",
                "latency",
                "tokens",
                "cost"
              ],
              "default": "timestamp"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "selected_key_ids",
            "in": "query",
            "description": "Comma-separated list of API key IDs to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "virtual_key_ids",
            "in": "query",
            "description": "Comma-separated list of virtual key IDs to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of logs matching the criteria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete logs",
        "description": "Deletes logs by their IDs.",
        "operationId": "deleteLogs",
        "tags": [
          "Logging"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of log IDs to delete"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logs deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - missing or empty IDs array"
          },
          "500": {
            "description": "Failed to delete logs"
          }
        }
      }
    },
    "/api/logs/stats": {
      "get": {
        "summary": "Get log statistics",
        "description": "Retrieves aggregated statistics for logs matching the filter criteria.",
        "operationId": "getLogsStats",
        "tags": [
          "Logging"
        ],
        "parameters": [
          {
            "name": "providers",
            "in": "query",
            "description": "Comma-separated list of providers to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "models",
            "in": "query",
            "description": "Comma-separated list of models to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Comma-separated list of statuses to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "Start time for the time range filter (RFC3339 format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "End time for the time range filter (RFC3339 format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogStats"
                }
              }
            }
          },
          "500": {
            "description": "Failed to calculate statistics"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatTool": {
        "type": "object",
        "required": [
          "type",
          "function"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "function"
            ],
            "description": "The type of the tool. Currently, only `function` is supported."
          },
          "function": {
            "type": "object",
            "required": [
              "name"
            ],
            "properties": {
              "description": {
                "type": "string",
                "description": "A description of what the function does."
              },
              "name": {
                "type": "string",
                "description": "The name of the function to be called."
              },
              "parameters": {
                "type": "object",
                "description": "The parameters the functions accepts, described as a JSON Schema object."
              }
            }
          }
        }
      },
...
      "ChatMessage": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "content": {
            "oneOf": [
              {
                "type": "string",
                "description": "Simple text content",
                "example": "Hello, how are you?"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ContentBlock"
                },
                "description": "Structured content with text and images"
              }
            ],
            "description": "Message content - can be simple text or structured content with text and images"
          },
          "tool_call_id": {
            "type": "string",
            "description": "ID of the tool call (for tool messages)"
          },
          "tool_calls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "description": "Tool calls made by assistant"
          },
          "refusal": {
            "type": "string",
            "description": "Refusal message from assistant"
          },
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            },
            "description": "Message annotations"
          },
          "thought": {
            "type": "string",
            "description": "Assistant's internal thought process"
          }
        }
      },
      "MessageRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant",
          "system",
          "tool"
        ],
        "description": "Role of the message sender",
        "example": "user"
      },
...
      "BifrostError": {
        "type": "object",
        "required": [
          "is_bifrost_error",
          "error"
        ],
        "properties": {
          "event_id": {
            "type": "string",
            "description": "Unique error event ID",
            "example": "evt_123"
          },
          "type": {
            "type": "string",
            "description": "Error type",
            "example": "invalid_request_error"
          },
          "is_bifrost_error": {
            "type": "boolean",
            "description": "Whether error originated from Bifrost",
            "example": true
          },
          "status_code": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 400
          },
          "error": {
            "$ref": "#/components/schemas/ErrorField"
          }
        }
      },
...
      "ResponsesMessage": {
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "content": {
            "$ref": "#/components/schemas/ResponsesMessageContent"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "LogStats": {
        "type": "object",
        "properties": {
          "total_requests": {
            "type": "integer"
          },
          "total_tokens": {
            "type": "integer"
          },
          "total_cost": {
            "type": "number"
          },
          "average_latency": {
            "type": "number"
          },
          "success_rate": {
            "type": "number"
          },
          "requests_by_provider": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "requests_by_model": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "requests_by_status": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        }
      },
...
      "Log": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "parent_request_id": { "type": "string", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" },
          "object": { "type": "string" },
          "provider": { "type": "string" },
          "model": { "type": "string" },
          "latency": { "type": "number", "format": "double", "nullable": true },
          "cost": { "type": "number", "format": "double", "nullable": true },
          "status": { "type": "string" },
          "stream": { "type": "boolean" },
          "raw_response": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "input_history": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChatMessage" }
          },
          "output_message": {
            "type": "object",
            "$ref": "#/components/schemas/ChatMessage",
            "nullable": true
          },
          "responses_output": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ResponsesMessage" }
          },
          "embedding_output": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EmbeddingData" }
          },
          "params": { "type": "object", "nullable": true },
          "tools": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChatTool" }
          },
          "tool_calls": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChatAssistantMessageToolCall" }
          },
          "token_usage": {
            "type": "object",
            "$ref": "#/components/schemas/BifrostLLMUsage",
            "nullable": true
          },
          "error_details": {
            "type": "object",
            "$ref": "#/components/schemas/BifrostError",
            "nullable": true
          },
          "speech_input": {
            "type": "object",
            "$ref": "#/components/schemas/SpeechInput",
            "nullable": true
          },
          "transcription_input": {
            "type": "object",
            "$ref": "#/components/schemas/TranscriptionInput",
            "nullable": true
          },
          "speech_output": {
            "type": "object",
            "$ref": "#/components/schemas/BifrostSpeechResponse",
            "nullable": true
          },
          "transcription_output": {
            "type": "object",
            "$ref": "#/components/schemas/BifrostTranscriptionResponse",
            "nullable": true
          },
          "cache_debug": {
            "type": "object",
            "$ref": "#/components/schemas/BifrostCacheDebug",
            "nullable": true
          }
        }
      },
      "EmbeddingData": {
        "type": "object",
        "properties": {
          "index": { "type": "integer" },
          "object": { "type": "string", "example": "embedding" },
          "embedding": { "$ref": "#/components/schemas/EmbeddingStruct" }
        }
      },
...
      "BifrostLLMUsage": {
        "type": "object",
        "properties": {
          "prompt_tokens": { "type": "integer" },
          "completion_tokens": { "type": "integer" },
          "total_tokens": { "type": "integer" }
        }
      },
      "SpeechInput": {
        "type": "object",
        "properties": {
          "input": { "type": "string" }
        }
      },
      "TranscriptionInput": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "byte",
            "description": "The audio file to transcribe."
          }
        }
      },
...
      "ChatAssistantMessageToolCall": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "function"
          },
          "id": {
            "type": "string"
          },
          "function": {
            "$ref": "#/components/schemas/ChatAssistantMessageToolCallFunction"
          }
        },
        "required": [
          "function"
        ]
      },
...
      "PaginationOptions": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "sort_by": {
            "type": "string"
          },
          "order": {
            "type": "string"
          }
        }
      },
      "SearchStats": {
        "type": "object",
        "properties": {
          "total_requests": {
            "type": "integer",
            "format": "int64"
          },
          "success_rate": {
            "type": "number",
            "format": "double"
          },
          "average_latency": {
            "type": "number",
            "format": "double"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_cost": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Log"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationOptions"
          },
          "stats": {
            "$ref": "#/components/schemas/SearchStats"
          }
        }
      },
...
    },
...
  }
}